phase-3_video-pipeline/requirements.txt
edge-tts>=7.0
moviepy>=2.0
Pillow>=10.0
requests>=2.30
python-dotenv>=1.0
